+++ ruby/recheck-rails/changelog.md
@@ -4,6 +4,14 @@
 
 * Public beta.
 
-## [0.1.0] - 2024-03-16
+* setup: Generate a query for LengthValidation.
+* setup: Generate a Checker from ActiveRecord validations.
+  Only PresenceValidation to start.
+
+## 0.4.0 - 2025-07-16
+
+* Public beta.
+
+## 0.1.0 - 2024-03-16
 
 * Placeholder.