+++ ruby/recheck/changelog.md
@@ -2,8 +2,13 @@
 
 ## [0.4.0] - 2025-07-16
 
+
+* Runner, reporter: add `around_query`
+
+## 0.4.0 - 2025-07-16
+
 * Public beta.
 
-## [0.1.0] - 2024-03-16
+## 0.1.0 - 2024-03-16
 
 * Placeholder.